bot/requirements.txt
discord.py==1.2.2
PyYAML==5.1.1
asyncpg==0.18.3
tortoise-orm==0.12.2 
